New UI Main page
Added:
Upload Status a bit less taller, to allow Run Analysis in the same page, no need scrolling
Add imputation level option for user (affects a lot)
Factor query samples to the last level in all plots
Add color choosing option, by batch, or tissue or diet 

Issue now is that colors and imputation are not saved for the server, I tried to do it but it ket bugging. I will keep working on it for the week.